Port over existing user tests to user-provider.test.ts
