+++ src/twitter-provider.test.ts
@@ -381,278 +381,6 @@ describe("TwitterProvider", () => {
     // #endregion getTweet
 
     // -----------------------------------------------------------------------------------------
-    // #region getUserByUsername
-    // -----------------------------------------------------------------------------------------
-
-    describe("getUserByUsername", () => {
-        test("given user exists, returns user", async () => {
-            // Arrange & Act
-            const username = USERNAME_BSCOTTORIGINALS;
-            const result = await TestTwitterProvider.getUserByUsername({
-                username,
-            });
-
-            // Assert
-            expect(result.data).toBeDefined();
-            expect(result.data!.username).toBe(username);
-        });
-
-        test("given user does not exist, returns undefined with errors", async () => {
-            // Arrange & Act
-            const username = fakeUsername();
-            const result = await TestTwitterProvider.getUserByUsername({
-                username,
-            });
-
-            // Assert
-            expect(result.data).toBeUndefined();
-            expect(result.errors?.length).toBeGreaterThanOrEqual(1);
-        });
-
-        test.each([
-            [UserFields.Verified, UserFields.CreatedAt],
-            `${UserFields.Verified},${UserFields.CreatedAt}`,
-        ])(
-            `when userFields %p and '${TweetExpansions.AuthorId}', returns additional fields`,
-            async (userFields) => {
-                // Arrange
-                const username = USERNAME_BSCOTTORIGINALS;
-
-                // Act
-                const result = await TestTwitterProvider.getUserByUsername({
-                    username,
-                    userFields,
-                });
-
-                // Assert
-                expect(result.data).toBeDefined();
-
-                const user = result.data!;
-
-                expect(user.username).toBeDefined();
-                expect(user.created_at).toBeDefined();
-                expect(user.verified).toBeDefined();
-            }
-        );
-
-        test.each([
-            [UserExpansions.PinnedTweetId],
-            `${UserExpansions.PinnedTweetId}`,
-        ])(
-            "when expansions %p, returns additional field",
-            async (expansions) => {
-                // Arrange
-                const username = USERNAME_TWITTERDEV;
-
-                // Act
-                const result = await TestTwitterProvider.getUserByUsername({
-                    username,
-                    expansions,
-                });
-
-                // Assert
-                expect(result.data).toBeDefined();
-
-                const user = result.data!;
-
-                expect(user.pinned_tweet_id).toBeDefined();
-            }
-        );
-
-        test.each([
-            [TweetFields.CreatedAt, TweetFields.Lang],
-            `${TweetFields.CreatedAt},${TweetFields.Lang}`,
-        ])(
-            `when fields %p and expansions ${UserExpansions.PinnedTweetId}, returns additional fields`,
-            async (fields) => {
-                // Arrange
-                const username = USERNAME_TWITTERDEV;
-
-                // Act
-                const result = await TestTwitterProvider.getUserByUsername({
-                    username,
-                    expansions: UserExpansions.PinnedTweetId,
-                    fields,
-                });
-
-                // Assert
-                expect(result.data).toBeDefined();
-
-                const user = result.data!;
-
-                expect(user.pinned_tweet_id).toBeDefined();
-                expect(result.includes?.tweets).toBeDefined();
-
-                const tweet = result.includes?.tweets![0]!;
-                expect(tweet.lang).toBeDefined();
-                expect(tweet.created_at).toBeDefined();
-            }
-        );
-
-        test(`when fields requested without expansions '${UserExpansions.PinnedTweetId}', returns additional fields`, async () => {
-            // Arrange
-            const username = USERNAME_TWITTERDEV;
-            const fields = [TweetFields.CreatedAt, TweetFields.Lang];
-
-            // Act
-            const result = await TestTwitterProvider.getUserByUsername({
-                username,
-                fields,
-            });
-
-            // Assert
-            expect(result.data).toBeDefined();
-
-            const user = result.data!;
-
-            expect(user.pinned_tweet_id).toBeDefined();
-            expect(result.includes?.tweets).toBeDefined();
-
-            const tweet = result.includes?.tweets![0]!;
-            expect(tweet.lang).toBeDefined();
-            expect(tweet.created_at).toBeDefined();
-        });
-    });
-
-    // #endregion getUserByUsername
-
-    // -----------------------------------------------------------------------------------------
-    // #region getUser
-    // -----------------------------------------------------------------------------------------
-
-    describe("getUser", () => {
-        test("given user exists, returns user", async () => {
-            // Arrange & Act
-            const id = USERID_BSCOTTORIGINALS;
-            const result = await TestTwitterProvider.getUser({
-                id,
-            });
-
-            // Assert
-            expect(result.data).toBeDefined();
-            expect(result.data!.id).toBe(id);
-        });
-
-        test("given user does not exist, returns undefined with errors", async () => {
-            // Arrange & Act
-            const id = "00000";
-            const result = await TestTwitterProvider.getUser({
-                id,
-            });
-
-            // Assert
-            expect(result.data).toBeUndefined();
-            expect(result.errors?.length).toBeGreaterThanOrEqual(1);
-        });
-
-        test.each([
-            [UserFields.Verified, UserFields.CreatedAt],
-            `${UserFields.Verified},${UserFields.CreatedAt}`,
-        ])(
-            `when userFields %p and '${TweetExpansions.AuthorId}', returns additional fields`,
-            async (userFields) => {
-                // Arrange
-                const id = USERID_BSCOTTORIGINALS;
-
-                // Act
-                const result = await TestTwitterProvider.getUser({
-                    id,
-                    userFields,
-                });
-
-                // Assert
-                expect(result.data).toBeDefined();
-
-                const user = result.data!;
-
-                expect(user.username).toBeDefined();
-                expect(user.created_at).toBeDefined();
-                expect(user.verified).toBeDefined();
-            }
-        );
-
-        test.each([
-            [UserExpansions.PinnedTweetId],
-            `${UserExpansions.PinnedTweetId}`,
-        ])(
-            "when expansions %p, returns additional field",
-            async (expansions) => {
-                // Arrange
-                const id = USERID_TWITTERDEV;
-
-                // Act
-                const result = await TestTwitterProvider.getUser({
-                    id,
-                    expansions,
-                });
-
-                // Assert
-                expect(result.data).toBeDefined();
-
-                const user = result.data!;
-
-                expect(user.pinned_tweet_id).toBeDefined();
-            }
-        );
-
-        test.each([
-            [TweetFields.CreatedAt, TweetFields.Lang],
-            `${TweetFields.CreatedAt},${TweetFields.Lang}`,
-        ])(
-            `when fields %p and expansions ${UserExpansions.PinnedTweetId}, returns additional fields`,
-            async (fields) => {
-                // Arrange
-                const id = USERID_TWITTERDEV;
-
-                // Act
-                const result = await TestTwitterProvider.getUser({
-                    id,
-                    expansions: UserExpansions.PinnedTweetId,
-                    fields,
-                });
-
-                // Assert
-                expect(result.data).toBeDefined();
-
-                const user = result.data!;
-
-                expect(user.pinned_tweet_id).toBeDefined();
-                expect(result.includes?.tweets).toBeDefined();
-
-                const tweet = result.includes?.tweets![0]!;
-                expect(tweet.lang).toBeDefined();
-                expect(tweet.created_at).toBeDefined();
-            }
-        );
-
-        test(`when fields requested without expansions '${UserExpansions.PinnedTweetId}', returns additional fields`, async () => {
-            // Arrange
-            const id = USERID_TWITTERDEV;
-            const fields = [TweetFields.CreatedAt, TweetFields.Lang];
-
-            // Act
-            const result = await TestTwitterProvider.getUser({
-                id,
-                fields,
-            });
-
-            // Assert
-            expect(result.data).toBeDefined();
-
-            const user = result.data!;
-
-            expect(user.pinned_tweet_id).toBeDefined();
-            expect(result.includes?.tweets).toBeDefined();
-
-            const tweet = result.includes?.tweets![0]!;
-            expect(tweet.lang).toBeDefined();
-            expect(tweet.created_at).toBeDefined();
-        });
-    });
-
-    // #endregion getUser
-
-    // -----------------------------------------------------------------------------------------
     // #region listMentionsByUser
     // -----------------------------------------------------------------------------------------
 
@@ -1108,55 +836,4 @@ describe("TwitterProvider", () => {
     });
 
     // #endregion listTweetsByUsername
-
-    // -----------------------------------------------------------------------------------------
-    // #region listUsersByUsername
-    // -----------------------------------------------------------------------------------------
-
-    describe("listUsersByUsername", () => {
-        testReturnsUsers<ListUsersByUsernameParams>({
-            name: "given a valid username, it returns a user",
-            method: (sut) => sut.listUsersByUsername,
-            params: { usernames: USERNAME_BRANDONSCOTT },
-        });
-
-        testReturnsUsers<ListUsersByUsernameParams>({
-            name: "given comma separated usernames, it returns users",
-            method: (sut) => sut.listUsersByUsername,
-            params: {
-                usernames: `${USERNAME_BRANDONSCOTT},${USERNAME_BSCOTTORIGINALS}`,
-            },
-        });
-
-        testReturnsUsers<ListUsersByUsernameParams>({
-            name: "given array of usernames, it returns users",
-            method: (sut) => sut.listUsersByUsername,
-            params: {
-                usernames: [USERNAME_BRANDONSCOTT, USERNAME_BSCOTTORIGINALS],
-            },
-        });
-
-        testUserExpansionsReturnsExpandedFields<ListUsersByUsernameParams>({
-            method: (sut) => sut.listUsersByUsername,
-            params: {
-                usernames: USERNAME_TWITTERDEV,
-            },
-        });
-
-        testUserFieldsReturnsFields<ListUsersByUsernameParams>({
-            method: (sut) => sut.listUsersByUsername,
-            params: {
-                usernames: USERNAME_TWITTERDEV,
-            },
-        });
-
-        testTweetFieldsReturnsFields<ListUsersByUsernameParams>({
-            method: (sut) => sut.listUsersByUsername,
-            params: {
-                usernames: USERNAME_TWITTERDEV,
-            },
-        });
-    });
-
-    // #endregion listUsersByUsername
 });
